# core/src/main/java/ru/university/assistant/iam/api/LoginRequest.java
package ru.university.assistant.iam.api;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(@NotBlank @Email String email, @NotBlank String password) {}